requirements.txt
python-dotenv
Flask
flask-cors
langchain-qdrant
langchain-openai
langchain
qdrant-client
langchain-core
secure-smtplib
openai